enhances component presentation
