+++ src/index.css
@@ -777,7 +777,7 @@ h5#trips-column {
 
 .highscore-div-main {
   text-align: center;
-  margin-top: 10%
+  margin-top: 10%;
 }
 
 .result-title-div {
